feat: add routing and responsive navbar
