+++ client/src/pages/About.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const About = () => (
+  <section style={{ padding: "2rem", textAlign: "center" }}>
+    <h1>About</h1>
+    <p>Learn more about Color Palette Vault.</p>
+  </section>
+);
+
+export default About;
+
