+++ client/src/pages/Login.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Login = () => (
+  <section style={{ padding: "2rem", textAlign: "center" }}>
+    <h1>Login</h1>
+  </section>
+);
+
+export default Login;
+
