+++ client/src/pages/Register.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Register = () => (
+  <section style={{ padding: "2rem", textAlign: "center" }}>
+    <h1>Register</h1>
+  </section>
+);
+
+export default Register;
+
